pyproject.toml
[project]
name = "startup-reviver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.2.4",
    "langchain>=1.0.3",
    "langchain-anthropic>=1.0.1",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.5",
    "perplexityai>=0.19.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=25.9.0",
    "isort>=7.0.0",
    "pre-commit>=4.3.0",
]
